+++ dessia_common/schemas/core.py
@@ -27,6 +27,25 @@ TYPES_FROM_STRING = {'unicode': str, 'str': str, 'float': float, 'int': int, 'bo
 _fullargsspec_cache = {}
 
 
+class UntypedArgument(FailedCheck):
+    """ Used when an argument is not typed in function or class definition. """
+
+    def __init__(self, attribute: str):
+        super().__init__(f"Attribute '{attribute}' : has no typing")
+
+
+class WrongNumberOfArguments(FailedCheck):
+    """ Used when a typing does not have the right amount of arguments. """
+
+
+class WrongType(FailedCheck):
+    """ Used when an annotation does not have the right type. """
+
+
+class UnsupportedDefault(FailedCheck):
+    """ Used when an argument defines a default that is not supported. """
+
+
 class Schema:
     """
     Abstraction of a Schema.
@@ -43,9 +62,10 @@ class Schema:
      We could translate inspect argspecs into a dessia_common pseudo-language.
     """
 
-    def __init__(self, annotations: Annotations, argspec: inspect.FullArgSpec, docstring: str):
+    def __init__(self, annotations: Annotations, argspec: inspect.FullArgSpec, docstring: str, name: str = ""):
         self.annotations = annotations
         self.attributes = [a for a in argspec.args if a not in RESERVED_ARGNAMES]
+        self.name = name
 
         try:  # Parse docstring
             self.parsed_docstring = parse_docstring(docstring=docstring, annotations=annotations)
@@ -54,12 +74,12 @@ class Schema:
 
         self.parsed_attributes = self.parsed_docstring['attributes']
 
-        self.required_arguments, self.default_arguments = split_default_args(argspecs=argspec, merge=False)
+        self.required_arguments, default_arguments = split_default_args(argspecs=argspec, merge=False)
 
         self.property_schemas = {}
         for attribute in self.attributes:
             if attribute in self.annotations:
-                default = self.default_arguments.get(attribute, None)
+                default = default_arguments.get(attribute, None)
                 annotation = self.annotations[attribute]
                 schema = get_schema(annotation=annotation, attribute=attribute, definition_default=default)
                 self.property_schemas[attribute] = schema
@@ -135,14 +155,15 @@ class Schema:
                 issues += schema.check_list()
         return issues
 
+    @property
     def is_valid(self) -> bool:
         """ Return whether the class definition is valid or not. """
-        return self.check_list().checks_above_level("error")
+        return not self.check_list().checks_above_level("error")
 
     def attribute_is_annotated(self, attribute: str) -> PassedCheck:
         """ Check whether given attribute is annotated in function definition or not. """
         if attribute not in self.annotations:
-            return FailedCheck(f"Attribute '{attribute}' : has no typing")
+            return UntypedArgument(attribute)
         return PassedCheck(f"Attribute '{attribute}' : is annotated")
 
 
@@ -162,7 +183,8 @@ class ClassSchema(Schema):
         members = inspect.getfullargspec(self.class_.__init__)
         docstring = class_.__doc__
 
-        Schema.__init__(self, annotations=annotations, argspec=members, docstring=docstring)
+        Schema.__init__(self, annotations=annotations, argspec=members, docstring=docstring,
+                        name=full_classname(class_))
 
     @property
     def editable_attributes(self):
@@ -185,14 +207,35 @@ class MethodSchema(Schema):
     """
 
     def __init__(self, method: Callable):
+        if isinstance(method, property):
+            method = method.fget
         self.method = method
 
         annotations = get_type_hints(method)
         members = inspect.getfullargspec(method)
+        self.return_annotation = annotations.get("return", None)
         docstring = method.__doc__
-        Schema.__init__(self, annotations=annotations, argspec=members, docstring=docstring)
+        super().__init__(annotations=annotations, argspec=members, docstring=docstring, name=method.__name__)
 
         self.required_arguments = [str(self.attributes.index(a)) for a in self.required_arguments]
+
+    @property
+    def serialized(self):
+        """ Get a dictionary with serialized annotations of method arguments. """
+        return {k: s.serialized for k, s in self.property_schemas.items()}
+
+    @property
+    def return_schema(self):
+        """ Get the schema of the return annotation. """
+        return get_schema(annotation=self.return_annotation, attribute="return")
+
+    @property
+    def return_serialized(self):
+        """ Get the serialized annotation of the return annotation. """
+        try:
+            return self.return_schema.serialized
+        except NotImplementedError:
+            return None
 
     def to_dict(self):
         """ Write the whole schema. """
@@ -201,6 +244,37 @@ class MethodSchema(Schema):
         schema.update({"required": self.required_arguments, "properties": properties,
                        "description": self.parsed_docstring["description"]})
         return schema
+
+    def definition_json(self):
+        """ Dictionary that denotes the good definition of the method. Useful for low-code features. """
+        return {"name": self.name, "return": self.return_serialized, "arguments": self.serialized,
+                "valid": self.is_valid, "checks": self.check_list().to_dict()["checks"]}
+
+    def check_list(self) -> CheckList:
+        """
+        Browse all properties and List potential issues.
+
+        Checks performed for each argument :
+        - Is typed in method definition
+        - Schema specific check
+        """
+        issues = super().check_list()
+        issues += CheckList([self.return_is_annotated(), self.return_type_is_valid()])
+        return issues
+
+    def return_is_annotated(self) -> PassedCheck:
+        """ Check whether method return is annotated in definition or not. """
+        if "return" not in self.annotations:
+            return CheckWarning("Method return : is not annotated")
+        return PassedCheck("Method return : is annotated")
+
+    def return_type_is_valid(self) -> PassedCheck:
+        """ Check whether given attribute is annotated in function definition or not. """
+        try:
+            _ = self.return_schema
+            return PassedCheck(f"Method return : '{self.return_annotation}' is valid")
+        except NotImplementedError:
+            return CheckWarning(f"Method return : '{self.return_annotation} is not valid")
 
 
 class Property:
@@ -321,7 +395,7 @@ class TypingProperty(Property):
             pretty_origin = prettyname(self.origin.__name__)
             msg = f"{self.check_prefix}is typed as a '{pretty_origin}' which requires exactly 1 argument. " \
                   f"Expected '{pretty_origin}[T]', got '{self.annotation}'."
-            return FailedCheck(msg)
+            return WrongNumberOfArguments(msg)
         return PassedCheck(f"{self.check_prefix}has exactly one arg in its definition.")
 
 
@@ -549,7 +623,7 @@ class CustomClass(Property):
     def is_dessia_object_typed(self) -> PassedCheck:
         """ Check whether if typing for given attribute annotates a subclass of DessiaObject or not . """
         if not issubclass(self.annotation, CoreDessiaObject):
-            return FailedCheck(f"{self.check_prefix}Class '{self.serialized}' is not a subclass of DessiaObject.")
+            return WrongType(f"{self.check_prefix}Class '{self.serialized}' is not a subclass of DessiaObject.")
         msg = f"{self.check_prefix}Class '{self.serialized}' is properly typed as a subclass of DessiaObject."
         return PassedCheck(msg)
 
@@ -610,7 +684,7 @@ class UnionProperty(TypingProperty):
             return PassedCheck(msg)
         msg = f"{self.check_prefix}'standalone_in_db' values for arguments of Union type '{self.annotation}'" \
               f"are not consistent. They should be all standalone in db or none of them should."
-        return FailedCheck(msg)
+        return WrongType(msg)
 
 
 class HeterogeneousSequence(TypingProperty):
@@ -681,7 +755,7 @@ class HeterogeneousSequence(TypingProperty):
         if len(self.args) == 0:
             msg = f"{self.check_prefix}is typed as a 'Tuple' which requires at least 1 argument. " \
                   f"Expected 'Tuple[T0, T1, ..., Tn]', got '{self.annotation}'."
-            return FailedCheck(msg)
+            return WrongNumberOfArguments(msg)
         return PassedCheck(f"{self.check_prefix}has at least one argument : '{self.annotation}'.")
 
     def ellipsis_has_exactly_two_args(self) -> PassedCheck:
@@ -693,7 +767,7 @@ class HeterogeneousSequence(TypingProperty):
         if self.additional_items and len(self.args) != 2:
             msg = f"{self.check_prefix}is typed as an ellipsed 'Tuple' which requires at exactaly 2 arguments. " \
                   f"Expected 'Tuple[T, ...]', got '{self.annotation}'."
-            return FailedCheck(msg)
+            return WrongNumberOfArguments(msg)
         return PassedCheck(f"{self.check_prefix}is not an ill-defined ellipsed tuple : '{self.annotation}'.")
 
 
@@ -736,7 +810,7 @@ class HomogeneousSequence(TypingProperty):
         if self.definition_default is not None:
             msg = f"{self.check_prefix}Mutable List input defines a default value other than None," \
                   f"which will lead to unexpected behavior and therefore, is not supported."
-            return FailedCheck(msg)
+            return UnsupportedDefault(msg)
         msg = f"{self.check_prefix}Mutable List doesn't define a default value other than None."
         return PassedCheck(msg)
 
@@ -791,7 +865,7 @@ class DynamicDict(TypingProperty):
         if len(self.args) != 2:
             msg = f"{self.check_prefix}is typed as a 'Dict' which requires exactly 2 arguments. " \
                   f"Expected 'Dict[KeyType, ValueType]', got '{self.annotation}'."
-            return FailedCheck(msg)
+            return WrongNumberOfArguments(msg)
         return PassedCheck(f"{self.check_prefix}has two args in its definition : '{self.annotation}'.")
 
     def has_string_keys(self):
@@ -801,7 +875,7 @@ class DynamicDict(TypingProperty):
             # Should we support other types ? Numeric ?
             msg = f"{self.check_prefix}is typed as a 'Dict[{key_type}, {value_type}]' " \
                   f"which requires str as its key type. Expected 'Dict[str, ValueType]', got '{self.annotation}'."
-            return FailedCheck(msg)
+            return WrongType(msg)
         return PassedCheck(f"{self.check_prefix}has str keys : '{self.annotation}'.")
 
     def has_simple_values(self):
@@ -811,7 +885,7 @@ class DynamicDict(TypingProperty):
             msg = f"{self.check_prefix}is typed as a 'Dict[{key_type}, {value_type}]' " \
                   f"which requires a builtin type as its value type. " \
                   f"Expected 'int', 'float', 'bool' or 'str', got '{value_type}'."
-            return FailedCheck(msg)
+            return WrongType(msg)
         return PassedCheck(f"{self.check_prefix}has simple values : '{self.annotation}'.")
 
     def has_no_default(self) -> PassedCheck:
@@ -819,7 +893,7 @@ class DynamicDict(TypingProperty):
         if self.definition_default is not None:
             msg = f"{self.check_prefix}Mutable Dict input defines a default value other than None," \
                   f"which will lead to unexpected behavior and therefore, is not supported."
-            return FailedCheck(msg)
+            return UnsupportedDefault(msg)
         msg = f"{self.check_prefix}Mutable Dict doesn't define a default value other than None."
         return PassedCheck(msg)
 
@@ -1004,6 +1078,19 @@ class GenericTypeProperty(Property):
         return chunk
 
 
+class AnyProperty(Property):
+    """ Handle Any typed (cannot be form inputs). """
+
+    def __init__(self, attribute: str):
+        super().__init__(annotation=Any, attribute=attribute, definition_default=None)
+
+    def to_dict(self, title: str = "", editable: bool = False, description: str = ""):
+        """ Write chunk of Any property. Useful for low-code features. """
+        chunk = super().to_dict(title=title, editable=editable, description=description)
+        chunk.update({"properties": {".*": ".*"}, "type": "object"})
+        return chunk
+
+
 def inspect_arguments(method: Callable, merge: bool = False):
     """ Wrapper around 'split_default_argument' method in order to call it from a method object. """
     method_full_name = f'{method.__module__}.{method.__qualname__}'
@@ -1059,6 +1146,8 @@ def get_schema(annotation: Type[T], attribute: str = "", definition_default: Opt
     if hasattr(annotation, '__origin__') and annotation.__origin__ is type:
         # Type is not considered a Typing as it has no arguments
         return ClassProperty(annotation=annotation, attribute=attribute, definition_default=definition_default)
+    if annotation is Any:
+        return AnyProperty(attribute)
     if inspect.isclass(annotation):
         return custom_class_schema(annotation=annotation, attribute=attribute, definition_default=definition_default)
     if isinstance(annotation, TypeVar):